+++ CHANGELOG.md
@@ -22,7 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fixed driver incompatibility with latest Windows Insider build
 - fixed SandMan File Panel treating registry hive log files such as `RegHive.LOG1` and `RegHive.LOG2` as descendants of `RegHive` because of their shared filename prefix, causing them to be omitted when deleting the selection together [#4788](https://github.com/sandboxie-plus/Sandboxie/issues/4788)
 
-
+- fixed ApiTrace stack exhaustion on Cygwin and other alternate stacks by asynchronously batching trace events on a dedicated worker thread
 
 ## [1.18.1 / 5.73.1] - 2026-07-26